move curve types into separate mod
