+++ risc0/bigint2/src/ec/mod.rs
@@ -22,62 +22,11 @@ use crate::ffi::{sys_bigint2_3, sys_bigint2_4};
 const ADD_BLOB: &[u8] = include_bytes_aligned!(4, "ec_add_256.blob");
 const DOUBLE_BLOB: &[u8] = include_bytes_aligned!(4, "ec_double_256.blob");
 
-/// The secp256k1 curve's prime as u32 digits, least significant digit first
-const SECP256K1_PRIME: [u32; EC_256_WIDTH_WORDS] = [
-    0xFFFFFC2F, 0xFFFFFFFE, 0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF,
-];
-const SECP256K1_CURVE: &WeierstrassCurve<EC_256_WIDTH_WORDS> =
-    &WeierstrassCurve::<EC_256_WIDTH_WORDS>::new(
-        SECP256K1_PRIME,
-        [0u32; EC_256_WIDTH_WORDS],
-        [7, 0, 0, 0, 0, 0, 0, 0],
-    );
-
 pub const EC_256_WIDTH_WORDS: usize = 256 / 32;
 
-/// Generic static curve configuration.
-pub trait Curve<const WIDTH: usize> {
-    const CURVE: &'static WeierstrassCurve<WIDTH>;
-}
+mod curve;
 
-/// An implementation of [Curve] for secp256k1.
-///
-/// This type should be used as a generic for [AffinePoint].
-#[derive(Debug, Eq, PartialEq, Ord, PartialOrd)]
-pub enum Secp256k1Curve {}
-
-impl Curve<EC_256_WIDTH_WORDS> for Secp256k1Curve {
-    const CURVE: &'static WeierstrassCurve<EC_256_WIDTH_WORDS> = SECP256K1_CURVE;
-}
-
-/// An elliptic curve over a prime field
-///
-/// The curve is given in short Weierstrass form y^2 = x^3 + ax + b. It supports a maximum `WIDTH` of its prime (and hence all coefficients and coordinates) given as number of 32-bit words (so the maximum bitwidth will be `32 * WIDTH`)
-#[derive(Debug, Eq, PartialEq)]
-pub struct WeierstrassCurve<const WIDTH: usize> {
-    // buffer[0] is the prime, buffer[1] is a, buffer[2] is b
-    buffer: [[u32; WIDTH]; 3],
-}
-
-impl<const WIDTH: usize> WeierstrassCurve<WIDTH> {
-    // TODO this constructor is prone to misuse, ideal to have named fields
-    pub const fn new(
-        prime: [u32; WIDTH],
-        a: [u32; WIDTH],
-        b: [u32; WIDTH],
-    ) -> WeierstrassCurve<WIDTH> {
-        WeierstrassCurve {
-            buffer: [prime, a, b],
-        }
-    }
-
-    /// The curve as concatenated u32s
-    ///
-    /// Little-endian, prime then a then b
-    fn as_u32s(&self) -> &[[u32; WIDTH]; 3] {
-        &self.buffer
-    }
-}
+pub use curve::{Curve, Secp256k1Curve, WeierstrassCurve};
 
 /// An affine point on an elliptic curve.
 #[derive(Debug, Eq, PartialEq)]
